Update extractFeatures
This is an example of how to use the metaData file to extract the features of a specific test.

In this example the Mobility Lab Walking test is used. However, if alternative tests such as Sway, TUG, or PULL are used the indices of features will need to be modified (such as on lines 25, 26, and 31). The initation size of the design matrix will also need to be changed (Line 6, p=108) to match the number of features exported from Mobility Lab